Refactored OverviewContent to component and function-call + fixed styling of dollar sign and coding tags
