+++ documentation-app/src/components/component-preview/component-preview.module.css
@@ -1,0 +1,37 @@
+.grid {
+  --ion-grid-padding: 0;
+}
+
+.card {
+  height: 100%;
+  padding: 0;
+}
+
+.cardHeader {
+  padding: 0;
+}
+
+.image {
+  background-size: cover;
+}
+
+.previewArea {
+  border-bottom: 1px solid var(--border-color);
+  height: 300px;
+  width: 100%;
+}
+
+.previewElement {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.title {
+  color: var(--ion-text-color);
+  padding: 1rem;
+}
+
+.teaserText {
+  color: var(--ion-text-color);
+}
