+++ documentation-app/src/components/page-content/overview-content/OverviewContent.tsx
@@ -1,28 +1,10 @@
-import styles from './overview-content.module.css';
+import ComponentPreview from '../../component-preview/ComponentPreview';
 import { IDocumentationPage } from '../../../helpers/structure';
-import { IonCard, IonCardContent, IonCardHeader, IonCardTitle, IonGrid, IonRow, IonCol } from '@ionic/react';
-import React from 'react';
 
-type IProps = {
-  componentPages: IDocumentationPage[];
-};
-
-const OverviewContent: React.FC<IProps> = ({ componentPages }) => (
-  <IonGrid>
-    <IonRow>
-      {[...componentPages, ...componentPages].map((c, i) => (
-        <IonCol key={i} className='mb-1' size='12' size-md='6' size-lg='6' size-xl='4'>
-          <IonCard key={i} className={`${styles.card} theme-border shadow-lg`}>
-            <IonCardHeader className={styles.cardHeader}>
-              {c.preview?.picture && !c.preview?.element && <div className={`${styles.image} ${styles.previewArea}`} style={{ backgroundImage: `url('${c.preview.picture}')` }} />}
-              {c.preview?.element && <div className={`${styles.previewArea} ${styles.previewElement} theme-bg`}>{c.preview.element}</div>}
-              <IonCardTitle className={styles.title}>{c.title}</IonCardTitle>
-            </IonCardHeader>
-            <IonCardContent className={styles.teaserText}>{c.preview?.description}</IonCardContent>
-          </IonCard>
-        </IonCol>
-      ))}
-    </IonRow>
-  </IonGrid>
+export const makeOverviewContent = (previews: IDocumentationPage[]): JSX.Element => (
+  <>
+    <h3>Some intro text</h3>
+    <ComponentPreview componentPages={previews} />
+    <h3>Some outro text</h3>
+  </>
 );
-export default OverviewContent;
