--- documentation-app/src/components/page-content/overview-content/overview-content.module.css
@@ -1,33 +1,0 @@
-.card {
-  height: 100%;
-  padding: 0;
-}
-
-.cardHeader {
-  padding: 0;
-}
-
-.image {
-  background-size: cover;
-}
-
-.previewArea {
-  border-bottom: 1px solid var(--border-color);
-  height: 300px;
-  width: 100%;
-}
-
-.previewElement {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.title {
-  color: var(--ion-text-color);
-  padding: 1rem;
-}
-
-.teaserText {
-  color: var(--ion-text-color);
-}
