+++ src/components/info-area/info-area.module.css
@@ -1,6 +1,6 @@
 .container {
   background: rgba(var(--ion-color-primary-rgb), 0.2);
-  border-left: 12px solid var(--ion-color-primary);
+  border-left: 12px solid rgba(var(--ion-color-primary-rgb), 0.6);
 }
 
 .container code {
