+++ styles/theme.css
@@ -25,6 +25,7 @@ th {
 code {
   font-family: 'Tomorrow', sans-serif;
   letter-spacing: 0.06rem;
+  line-height: 2rem;
 }
 
 p,
